cleanup header through. tsx
